+++ processor/src/test/resources/expected/com/example/arez/React4j_KeepAliveMemoizeArezReactComponent.java
@@ -4,6 +4,7 @@ import arez.Arez;
 import arez.Disposable;
 import arez.annotations.ArezComponent;
 import arez.annotations.Executor;
+import arez.annotations.Feature;
 import arez.annotations.Memoize;
 import arez.annotations.Observe;
 import arez.annotations.Priority;
@@ -23,6 +24,7 @@ import react4j.ReactNode;
 
 @ArezComponent(
     name = "KeepAliveMemoizeArezReactComponent",
+    disposeTrackable = Feature.DISABLE,
     deferSchedule = true
 )
 @Generated("react4j.processor.ReactProcessor")