+++ processor/src/test/resources/expected/com/example/arez/memoize/React4j_ParameterizedMemoizeComponent.java
@@ -5,6 +5,7 @@ import arez.Disposable;
 import arez.annotations.ArezComponent;
 import arez.annotations.DepType;
 import arez.annotations.Executor;
+import arez.annotations.Feature;
 import arez.annotations.Memoize;
 import arez.annotations.Observe;
 import arez.annotations.Priority;
@@ -24,6 +25,7 @@ import react4j.ReactNode;
 
 @ArezComponent(
     name = "ParameterizedMemoizeComponent",
+    disposeTrackable = Feature.DISABLE,
     deferSchedule = true
 )
 @Generated("react4j.processor.ReactProcessor")